:recycle: Split repository_selector into three files (view, model, viewmodel)
